+++ backend/profiling/trace_sources/trace_parser.cpp
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Intel Corporation
+ * Copyright (c) 2021-2025 Intel Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -317,7 +317,7 @@ void HpuTraceParser::convertEventsToActivities(
   };
   using ActiveEventsMap = std::unordered_map<
       uint32_t,
-      std::unordered_map<uint32_t, std::list<ActiveEvent>>>;
+      std::unordered_map<uint64_t, std::list<ActiveEvent>>>;
   using ActiveEnqueueEventsMap = std::unordered_map<uint32_t, synTraceEvent*>;
   ActiveEventsMap activeEvents;
   ActiveEnqueueEventsMap activeEnqueueEvents;
